# storage/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Library API'
  description: 'REST API for the library system (versioned under /api/v1).'
  version: 1.0.0
servers:
  -
    url: /api/v1
    description: 'Version 1'
paths:
  /books:
    get:
      tags:
        - Book
      summary: 'List and search books'
      operationId: booksIndex
      parameters:
        -
          name: title
          in: query
          required: false
          schema:
            type: string
        -
          name: author
          in: query
          required: false
          schema:
            type: string
        -
          name: genre
          in: query
          required: false
          schema:
            type: string
        -
          name: available_only
          in: query
          required: false
          schema:
            type: boolean
          example: true
        -
          name: sort_by
          in: query
          required: false
          schema:
            type: string
            default: title
            enum:
              - title
              - author
              - genre
              - created_at
              - available_copies
              - total_copies
        -
          name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 15
            maximum: 100
            minimum: 1
      responses:
        '200':
          description: 'Paginated books'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBooksResponse'
        '401':
          description: Unauthenticated
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
    post:
      tags:
        - Book
      summary: 'Create book'
      operationId: booksStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreBookRequestBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDataResponse'
        '401':
          description: Unauthenticated
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
  '/books/{book}':
    get:
      tags:
        - Book
      summary: 'Get book by id'
      operationId: booksShow
      parameters:
        -
          name: book
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDataResponse'
        '401':
          description: Unauthenticated
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Book
      summary: 'Delete book (forbidden if active rentals exist)'
      operationId: booksDestroy
      parameters:
        -
          name: book
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthenticated
        '404':
          description: 'Not found'
        '409':
          description: 'Active rentals'
      security:
        -
          sanctum: []
    patch:
      tags:
        - Book
      summary: 'Update book'
      operationId: booksUpdate
      parameters:
        -
          name: book
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequestBody'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDataResponse'
        '401':
          description: Unauthenticated
        '404':
          description: 'Not found'
        '409':
          description: Conflict
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
...
components:
  schemas:
    BookDataResponse:
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/BookResource'
        message:
          type: string
          example: 'Book updated'
      type: object
    BookResource:
      required:
        - id
        - title
        - author
        - genre
        - total_copies
        - available_copies
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        description:
          type: string
          nullable: true
        total_copies:
          type: integer
        available_copies:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    PaginatedBooksResponse:
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookResource'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
    StoreBookRequestBody:
      required:
        - title
        - author
        - genre
        - total_copies
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        description:
          type: string
          nullable: true
        total_copies:
          type: integer
          minimum: 0
        available_copies:
          type: integer
          minimum: 0
      type: object
    UpdateBookRequestBody:
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        description:
          type: string
          nullable: true
        total_copies:
          type: integer
          minimum: 0
        available_copies:
          type: integer
          minimum: 0
      type: object
...
    MessageResponse:
      required:
        - message
      properties:
        message:
          type: string
          example: 'Book deleted'
      type: object
    PaginationMeta:
      required:
        - current_page
        - per_page
        - total
        - last_page
      properties:
        current_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        last_page:
          type: integer
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Laravel Sanctum personal access token: Authorization: Bearer {token}'
      scheme: bearer
...